Implemented better pattern matching
Previously the "optimal pattern" was found by picking the three objects
with the best matching back-projection - but the back projection average
function was always 0, so it was essentially picking three at random.
Implemented a better method, which finds a center block with two outer
blocks touching it, one on either side.