Beginnings of SIFT code
This code looks for one out of any number of images provided by the user
using SIFT. This version functions but is not yet integrated with any of
the drone code (which is the goal). Committing now partly to have a
record of older code to refer to. Note also that a 'door' reference
image is included and can be used, but is much more inaccurate than the
other two.